factored out links, rotation handles, scale handles
