+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/BoundingBoxHelper.cs
@@ -1,5 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
+using Microsoft.MixedReality.Toolkit.Input;
+using Microsoft.MixedReality.Toolkit.UI.BoundingBoxTypes;
 using System.Collections.Generic;
 using UnityEngine;
 
@@ -233,61 +235,6 @@ namespace Microsoft.MixedReality.Toolkit.UI
             Vector3[] corners = null;
             cloneBounds.GetCornerPositions(ref corners);
             boundsPoints.AddRange(corners);
-        }
-
-
-        public static Bounds GetMaxBounds(GameObject g)
-        {
-            var b = new Bounds();
-            Mesh currentMesh;
-            foreach (MeshFilter r in g.GetComponentsInChildren<MeshFilter>())
-            {
-                if ((currentMesh = r.sharedMesh) == null) { continue; }
-
-                if (b.size == Vector3.zero)
-                {
-                    b = currentMesh.bounds;
-                }
-                else
-                {
-                    b.Encapsulate(currentMesh.bounds);
-                }
-            }
-            return b;
-        }
-
-
-
-        /// <summary>
-        /// Add all common components to a corner or rotate affordance
-        /// </summary>
-        /// <param name="afford"></param>
-        /// <param name="bounds"></param>
-        public static void AddComponentsToAffordance(GameObject afford, Bounds bounds, RotationHandlePrefabCollider colliderType, CursorContextInfo.CursorAction cursorType, Vector3 colliderPadding)
-        {
-            if (colliderType == RotationHandlePrefabCollider.Box)
-            {
-                BoxCollider collider = afford.AddComponent<BoxCollider>();
-                collider.size = bounds.size;
-                collider.center = bounds.center;
-                collider.size += colliderPadding;
-            }
-            else
-            {
-                SphereCollider sphere = afford.AddComponent<SphereCollider>();
-                sphere.center = bounds.center;
-                sphere.radius = bounds.extents.x;
-                sphere.radius += Mathf.Max(Mathf.Max(colliderPadding.x, colliderPadding.y), colliderPadding.z);
-            }
-
-            // In order for the affordance to be grabbed using near interaction we need
-            // to add NearInteractionGrabbable;
-            var g = afford.EnsureComponent<NearInteractionGrabbable>();
-            g.ShowTetherWhenManipulating = drawTetherWhenManipulating;
-
-            var contextInfo = afford.EnsureComponent<CursorContextInfo>();
-            contextInfo.CurrentCursorAction = cursorType;
-            contextInfo.ObjectCenter = rigRoot.transform;
         }
     }
 }
