+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals/BoundingBoxHandlesBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dbd3ad2db9a1de14fab93ba923124e5d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
