+++ Assets/MixedRealityToolkit.SDK/Features/UX/Scripts/BoundingBox/Visuals/BoundingBoxRotationHandles.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b4f699f84973b2c43b9953128882a76d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
